Update with a couple of new tips.
I've been working on STM32H7 and found the documentation very vague. A common pattern I've noticed is placing information directly relevant to one peripheral/system in the section of another. They don't seem to believe in the value of cross referencing. I've added a suggestion to search the whole document that can help to mitigate this.